Rename umh_stream to uns_plugin
