+++ cmd/benthos/bundle/package.go
@@ -28,5 +28,5 @@ import (
 	_ "github.com/united-manufacturing-hub/benthos-umh/s7comm_plugin"
 	_ "github.com/united-manufacturing-hub/benthos-umh/sensorconnect_plugin"
 	_ "github.com/united-manufacturing-hub/benthos-umh/tag_processor_plugin"
-	_ "github.com/united-manufacturing-hub/benthos-umh/umh_stream_plugin"
+	_ "github.com/united-manufacturing-hub/benthos-umh/uns_plugin"
 )
