+++ uns_plugin/uns_output.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package umh_stream_plugin
+package uns_plugin
 
 import (
 	"context"
